chore (refs T22614): move DpResettableInput to client/js.
