+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/SegmentsList/FilterFlyout.vue
@@ -161,7 +161,7 @@ import { DpButton, DpLoading } from 'demosplan-ui/components'
 import { mapGetters, mapMutations } from 'vuex'
 import dataTableSearch from '@DpJs/components/core/DpDataTable/DataTableSearch'
 import DpFlyout from '@DpJs/components/core/DpFlyout'
-import DpResettableInput from '@DemosPlanCoreBundle/components/DpResettableInput'
+import DpResettableInput from '@DpJs/components/core/DpResettableInput'
 import FilterFlyoutCheckbox from './FilterFlyoutCheckbox'
 import hasOwnProp from '@DpJs/lib/utils/hasOwnProp'
