+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/splitStatement/DpCreateTag.vue
@@ -69,7 +69,7 @@ import { DpInput, DpLabel } from 'demosplan-ui/components'
 import { mapActions, mapGetters, mapMutations } from 'vuex'
 import DpButtonRow from '@DpJs/components/core/DpButtonRow'
 import DpMultiselect from '@DpJs/components/core/form/DpMultiselect'
-import DpResettableInput from '@DemosPlanCoreBundle/components/DpResettableInput'
+import DpResettableInput from '@DpJs/components/core/DpResettableInput'
 import dpValidateMixin from '@DpJs/lib/validation/dpValidateMixin'
 
 export default {
